+++ docs/conf.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# zope2.signedsessioncookie documentation build configuration file, created by
+# zope2.sessioncookie documentation build configuration file, created by
 # sphinx-quickstart on Wed Sep 16 15:21:41 2015.
 #
 # This file is execfile()d with the current directory set to its
@@ -51,9 +51,9 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'zope2.signedsessioncookie'
-copyright = u'2015, ZeOmega, Inc.'
-author = u'ZeOmega, Inc.'
+project = u'zope2.sessioncookie'
+copyright = u'2015, Zope Foundation and Contributors.'
+author = u'Zope Foundation and Contributors.'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -205,7 +205,7 @@ html_static_path = ['_static']
 #html_search_scorer = 'scorer.js'
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'zope2signedsessioncookiedoc'
+htmlhelp_basename = 'zope2sessioncookiedoc'
 
 # -- Options for LaTeX output ---------------------------------------------
 
@@ -227,8 +227,8 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-  (master_doc, 'zope2signedsessioncookie.tex', u'zope2.signedsessioncookie Documentation',
-   u'ZeOmega, Inc.', 'manual'),
+  (master_doc, 'zope2sessioncookie.tex', u'zope2.sessioncookie Documentation',
+   u'Zope Foundation and Contributors.', 'manual'),
 ]
 
 # The name of an image file (relative to this directory) to place at the top of
@@ -257,7 +257,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'zope2signedsessioncookie', u'zope2.signedsessioncookie Documentation',
+    (master_doc, 'zope2sessioncookie', u'zope2.sessioncookie Documentation',
      [author], 1)
 ]
 
@@ -271,8 +271,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-  (master_doc, 'zope2signedsessioncookie', u'zope2.signedsessioncookie Documentation',
-   author, 'zope2signedsessioncookie', 'One line description of project.',
+  (master_doc, 'zope2sessioncookie', u'zope2.sessioncookie Documentation',
+   author, 'zope2sessioncookie', 'One line description of project.',
    'Miscellaneous'),
 ]