+++ zope2/sessioncookie/tests/test_config.py
@@ -52,13 +52,13 @@ class SignedSessionCookieConfigTests(unittest.TestCase):
         self.assertEqual(config.reissue_time, 234)
 
     def test_ctor_no_pyramid_nacl_session(self):
-        from zope2.signedsessioncookie import config as MUT
+        from zope2.sessioncookie import config as MUT
         with _Monkey(MUT, _HAS_PYRAMID_NACL_SESSION=False):
             with self.assertRaises(ValueError):
                 self._makeOne('SECRET', encrypt=True)
 
     def test_ctor_w_pycrypto(self):
-        from zope2.signedsessioncookie import config as MUT
+        from zope2.sessioncookie import config as MUT
         with _Monkey(MUT, _HAS_PYRAMID_NACL_SESSION=True):
             config = self._makeOne('SECRET', encrypt=True)
             self.assertTrue(config.encrypt)